Improved support for specifying prerequisites. Now the 'prerequisites' DSL allows plugins and build script to register prerequisite types, specify prerequisites (i.e. supplying any parameters), and check the prerequisites. Since all prerequisites have to be specified (even if they aren't eagerly checked) this allows a 'checkPrerequisites' task to go through them all when the user chooses. Failing that, the prerequisites should be checked lazily i.e. prerequisite failures shouldn't occur unless they need to occur to prevent real error conditions.
The intrepid plugin now has a dependency on selenic selenic:Mercurial:2.5.1, which has been published to Artifactory, also bundling up mercurial_keyring.py. This breaks the dependency on hg being available on the command line and allows the plugin to satisfy its own dependency. Also, since the mercurial.ini that's part of the selenic:Mercurial:2.5.1 package applies the mercurial_keyring extension by default, that simplifies that dependency too. However it is still necessary for an [auth] section to exist in the user's own mercurial.ini, and the intrepid plugin now checks for this explicitly. There is also a new task 'fixMercurialIni' which will fix up the user's mercurial.ini to include the [auth] section, because it's better to avoid changing global configuration on the user's machine without consent.