GR #2080 Bugfix
The artifactoryAPI.getFolderInfoJson() method
was using 'JsonSlurper' to parse the already-parsed JSON data
received by the RESTClient via HTTP, and therefore throwing an
exception.  The solution is to remove this superfluous code and
the plugin behaves as expected.

I suspect this was an accidental commit during debugging, pre-dating
the use of RESTClient.parsers which is what is doing the JSON parsing.