GR #2037 Bugfix: transitive dependencies across projects
It was possible to set up a multi-project environment,i.e., using
multiple "sourceDependency"s with their own gradle scripts, where
the transitive dependency of one project (in a "sourceDependency") conflicted
with the transitive dependency of another.

This was because gradle had not been told about the dependencies between
a project and its "sourceDependency" projects, so gradle did not see the
conflict.

The fix involves:
a) Adding the "packedDependency"s earlier (during script handling rather than
"ProjectsEvaluated" - which was too late for gradle's conflict management)

b) Adding dependencies between a project and each of its sourceDependency
projects