Build/release clean-ups.
* Added some static type info.
* Replaced the amusing but unhelpful "Beware the daemon" error message.
* Changed the default version for new plugins to have only three digits, as
that's our new recommended numbering scheme.
* Changed choosePluginVersion to take advantage of the recent addition of
project dependencies between plugins.  If we're publishing a module which
depends on some other module that we're also publishing, use the to-be-published
version number, instead of trying to look up the last published version.  This
makes it easy to publish multiple plugins at once.