GR #2281 Bugfixes
* Casting magic GPathResult objects to NodeChild doesn't always work.  Sometimes
they are NodeChildren objects, and that class isn't public in Gradle 1.4, and
doesn't have a common base with NodeChild which has the "text()" method -- it's
not on GPathResult.  So, we'll have to stick with dynamic-ness.

* IntelliJ claimed that "pomXml.parent" was an access to a non-visible member,
which I thought was possible but bad practice in Groovy.  So, in 8e6816b546db,
I changed it to a call to "pomXml.parent()", which returns the protected
property.  Except IntelliJ was wrong; it fact it was accessing a magic GPath
property to get the "<parent />" node from the XML.  Calling "parent()" returned
the XML for the entire POM file, not the POM file's "<parent />" element, and
the top-level file sometimes contains the same set of elements as a "<parent />",
but (of course) describing that very POM, not its parent.  This results in stack
overflow.