GR #2281 Rework
SourceDependencyHandler shouldn't default the "group" for a project.  I'm pretty
sure Scott did this as a workaround for a project which didn't publish anything
(because it was just pulled in as a sourceDependency for direct access to the
source), and hence didn't declare its group, and so things depending on it got
the wrong group.

However, (a) it could still declare its group anyway; and (b) the things
depending on it shouldn't really need to publish a dependency on it.  Intrepid
automatically forces projects to have a published dependency on their
sourceDependencies (something we should fix later), but we can work around this
by adding the following code in the depending project.

publishing {
    publications {
        ivy.descriptor.withXml { xml ->
            xml.asNode().dependencies.dependency.each { depNode ->
                if (depNode.@name == "projectB") {
                    depNode.parent().remove(depNode)
                }
            }
        }
    }
}