GR #2504 Review changes
I eventually worked out that Scott's change for this was fine, but I was puzzled
by DependencyHandler#getTargetName.  I eventually worked out it was trying to
return the last (directory) part of a path, using hand-crufted regexp parsing,
so I just changed it to use the Java File class.  I then realised that that bean
property method could be changed to a final field initialized in the constructor,
so I did that.