GR #2827 Add sitmap "pane", and "buildWebsiteQuickly" task.
The sitemap will be included on the left of every page.  This involved some CSS
changes and additions.

The buildWebsiteQuickly task uses no source highlighting, since pygments seems
to be slow.

Also, I found that MS-DOS "shift" splits arguments at "=", so had to put a hack
into the batch files for that.

Lastly, I renamed a couple of files.