GR #2144 Allow publishing to a local repo, as a step towards proper integration tests.
Some of the JUnit tests inherit from holygradle.test.TestBase, which launches Gradle to run what are
really integration tests.  However, they currently pull together the plugins required by manually
adding them to the classpath and setting properties which should have been set by the custom-gradle
init script.  Even that isn't enough in some cases, though; e.g., when credential-store.exe is
needed.  So, I'm changing the way publishing is done, to allow publishing of both the plugins and
the custom-gradle distribution to a local repo, so that integration tests can run using that, more
closely reproducing real usage.

The planned steps when publishing are as follows, though some of them may be tied together
automatically through task dependencies.
* compile
* run unit tests
* publish to local repo (release or snapshot version)
* run integration tests
* publish to real repo (release or snapshot version)

This commit changes the publishing process to do that, but doesn't yet change the unit tests.

To make the process simpler, I've changed the publishing process so that all plugins must now be
published together, all the time, with the same version number.  This was always a better idea,
really, since they all depend on having the same version of custom-gradle-core-plugin, even if they
don't (statically) depend on each other.  The custom-gradle distribution can still be published
separately.

This change also means that we can just use normal Gradle project dependencies between plugin
projects, rather than having special code in the Helper class to manage it.  And that means that we
can be sure that we always build and test things in the right order, and that, when you publish a
plugin, you're definitely publishing it with dependencies on other plugins which really match the
source (of those other plugins) against which it was built.  (Previously it would have been possible
to publish just intrepid, say, and it would publish a dependency on the latest published version of
custom-gradle-core-plugin.  If you had changed the source of custom-gradle-core-plugin when you
built and published intrepid, and forgot to also publish custom-gradle-core-plugin, you'd end up
with a published set of modules that didn't match any tested combination of things built together
from source.  Strictly speaking, you can still publish just a subset of the plugins together, with
one version number, but there won't be any versions of the other plugins which will be compatible
with them.)

File changes follow ...

* developer-readme.md
** Document changes to the publishing process.

* .hgignore
** Ignore the "local_repo" directory to which things will be published.

* build.gradle
** Change publishing so that each of "the custom-gradle distribution" and "all the plugins" can be
published to a local repo or a remote repo.  In each case, the version will be "<user>-SNAPSHOT"
unless the property "pubAs" is defined as "release", in which case the user will be prompted on the
console for a version string.
** For publishing to the local repo, use task intputs/outputs to avoid republishing if the artifacts
haven't changed.
** Use Configuration#getTaskDependencyFromProjectDependency so that publish tasks in different
projects depend on each other as appropriate.
** In passing, move the generated "ivy.xml" for each project into the "build" directory of that
project, instead of dumping all of them to the same top-level location.
** Add the local repo to the set of repos from which dependencies can be fetched.
** Make both "publishXxxReally" tasks depend on testing.
** Add comments marking where sections start/end.  I think I'll soon pull these out into separate
files.

** Just always ask the user to provide the version number when publishing a release, without doing
magic with previously published versions, to keep things simple.

* buildSrc\src\main\groovy\holygradle\util\Helper.groovy
** Now that we ask the user for release version info in a simple way, we don't need this class.

* credential-store\build.gradle
** To speed up build, add task inputs/outputs info, so that we don't run devenv.exe unnecessarily.
** Handle the fact that there are multiple publish tasks now.

* custom-gradle\build.gradle
** Handle the fact that there are multiple publish tasks now.

* devenv-plugin\build.gradle
* intrepid-plugin\build.gradle
* my-credentials-plugin\build.gradle
* stamper-plugin\build.gradle
* unit-test-plugin\build.gradle
** Use normal project dependencies instead of custom code, which is possible now that we always just
ask for the release version once, and apply it to all projects.

* settings.gradle
** Remove mention of wrapper-starter-kit, which was deleted from the repo some time ago.