GR #2614 Minor improvement to message on credentials dialog timeout.
The reported problem was that, when a sourceDependencies block pointed to a repo
for which no credentials were cached, the dialog requesting credentials didn't
appear.  A workaround seemed to be to run the build with "--no-daemon".

I now can't reproduce this problem.  The dialog appears for me, whether the
build is run normally from the console, or as part of an integration test using
BuildLauncher.  So, I've changed the error message to suggest using
"--no-daemon" but won't make any further changes.