GR #2144 Fix version stamping and publishing of credential-store.
Previously it wasn't always re-built or re-published when "gw.bat" was run with
a different version number from previous runs.

* credential-store\build.gradle
** The default version is now "<user>-SNAPSHOT", not "unspecified".

* publish.gradle
** Since the version number for "real" publishing is now passed on the command
line, instead of coming from an interactive query, we can set the project
versions earlier than "project.gradle.taskGraph.whenReady".  This means we can
have a task to write it to a file and make the
":credential-store:createReleaseVersionFile" task depend on that task.  This
means that credential-store.exe is guaranteed to be rebuilt if the version
changes.  This is the same mechanism used to make it rebuild when the hg hash
of the working directory changes.
** While debugging, noticed that the inputs of the publishPluginsLocally tasks
were including all JARs leftover from previous publish runs in the root
directory, not just the one for the current version, so I tidied that up.

* .hgignore
** Removed hgVersionInfo.txt since it's now in an already-ignored folder.

* versionInfo.gradle
** Move the "hgVersionInfo.txt" file to be in the project's "build" dir, for
tidiness.  (The new task for credential-store does the same.)
** Tidied up the calls to methods on setHgVersionInfo.outputs.