GR #2081 Fix problem with pinned sourceDependencies in meta-packages.
I reworked a change to SourceControlRepositories#get and incorrectly ended up
pulling out the url@node of the parent project, not the source depdendency.  I
didn't spot this in basic tests, because it only affects meta-package generation,
which failed in integration tests of projects using the plugins.

* intrepid-plugin.iml
** Remove javahg info from IntelliJ lists.

* src\main\groovy\holygradle\Helper.groovy
** Reads of extension properties should avoid going through the ".ext" property,
to take advantage of the Project property resolution mechanism.  See
<http://www.gradle.org/docs/1.4/dsl/org.gradle.api.Project.html#org.gradle.api.Project.properties>.
(In this case, a property has been moved from each project to existing just once
on the root project.)

* src\main\groovy\holygradle\IntrepidPlugin.groovy
** Made build-wide project properties live on the root project, not the current
project.  This is to make it clearer that it's okay for
SourceControlRepositories#get to just take the rootProject as an argument,
which is part of the changes for this fix.  Really these could/should be
convention properties, but that's for another day.

* src\main\groovy\holygradle\packaging\PackageArtifactBuildScriptHandler.groovy
** Sort-of revert the changes made in 4fc856116dd7; see comment for
SourceControlRepositories.

* src\main\groovy\holygradle\packaging\PackageArtifactHandler.groovy
** Follow interface change.

* src\main\groovy\holygradle\scm\SourceControlRepositories.groovy
** Now takes the rootProject (to get build script dependecies for the Hg case),
plus the path to the repo we're interested in (which may not belong to any
Gradle project).

* src\main\groovy\holygradle\scm\SvnDependency.groovy
** Reads of extension properties should avoid going through the ".ext" property.

* src\main\groovy\holygradle\source_dependencies\SourceDependencyHandler.groovy
** Reads of extension properties should avoid going through the ".ext" property.

* src\main\groovy\holygradle\unpacking\UnpackModule.groovy
** Reads of extension properties should avoid going through the ".ext" property.

* src\test\groovy\holygradle\scm\SourceControlRepositoriesTest.groovy
** Follow interface change.

* src\test\groovy\holygradle\unpacking\UnpackModuleVersionTest.groovy
** Reads of extension properties should avoid going through the ".ext" property.