GR #2070, GR #2081 Make sure Hg INI file is updated with mercurial_keyring path.
Intrepid unpacks its own copy of Mercurial, and then modifies the mercurial.ini
file to point to a bundled copy of the mercurial_keyring.py extensions.  Or at
least, it used to.  In 527de19e52eb (rev 89 for me), Neil made a change to
BuildScriptDependency.groovy which accidentally meant that this adjustment to
the INI file no longer happens.  The INI file update happens in a doLast block
added to the Mercurial unpack task in IntrepidPlugin#apply.  Originally
BuildScriptDependency created this unpack task as

  unpackTask = project.task(..., type: Copy) { ... }

Neil changed that to

  unpackTask = project.task(..., type: DefaultTask) {
     ...
     if (!exists ...) {
       project.copy(...)
     }
     ...
   }

That means that the copy (the unpack) happens when the task is defined, not when
it is executed; but the task, being a default task with no inputs or outputs
defined, never gets executed, so the INI file adjustment doesn't happen.

This problem only appears if you've had the intrepid plugin download Mercurial
since that problem was introduced.  I saw it recently when I started adding
integration tests, because they're using a custom local Gradle user home folder.

Reverting the task to be a Copy task fixes the problem.

* .hgignore
** The PackageArtifactBuildScriptHandlerIntegrationTest now runs "fAD" to better
simulate real used (in particular, to trigger the Mercurial unpack task), so we
need to ignore the settings file that generates.

* intrepid-plugin\src\main\groovy\holygradle\buildscript\BuildScriptDependency.groovy
** Change the normal unpackTask (back) to being a Copy task.  The guard of "if
the folder already exists" I moved to an onlyIf block.

* intrepid-plugin\src\main\groovy\holygradle\IntrepidPlugin.groovy
** While investigating this, I realised we don't want to be using the user's
main mercurial.ini file.  We switched to using this for all hg.exe calls as part
of #2081, but previously it was only every used for the "update" after an
initial clone -- I suppose it might have been a left-over from some changes
between using hg.exe and javahg.  We in fact definitely want to use the INI file
in the same folder as the unpacked hg.exe, because that gets modified by the
intrepid plugin after it's downloaded and unpacked.

* intrepid-plugin\src\main\groovy\holygradle\scm\HgCommandLine.groovy
* intrepid-plugin\src\main\groovy\holygradle\scm\SourceControlRepositories.groovy
* intrepid-plugin\src\main\groovy\holygradle\source_dependencies\SourceDependencyHandler.groovy
** Don't use the user's main mercurial.ini file.

* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest.groovy
* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest\projectAInput\setupRepo.gradle
* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest\projectAInput\build.gradle
** Run "fetchAllDependencies" first to better simulate real use (which would
trigger the unpackMercurial task).
** Because I'm running fetchAllDependencies, projectA is treated as a Gradle
subproject, so its gradle file needs to have the right name.
** Add an explicit task in the repo setup, instead of just running hg directly
during project evaluation, so it, too, can have a dependency on unpackMercurial.