GR #3723 intrepid may not produce sufficient symlink/unpack tasks for depedencies
The key change is UnpackModuleVersion class may now have multiple parents,
e.g. be a packed dependency as well as a transitive dependency of one or
more other packed dependencies.

shouldUnpackToCache() method modified to take 'multiple parents' into account,
as well as asserting that the unpack policy (to cache/to workspace) is
consistent.

getUnpackTasks() and getSymlinkTaskIfUnpackingToCache() can both now return
a set of tasks, i.e. in the case where the dependency needs to occur in
multiple places.  Of course, such structure shouldn't be encouraged, but
since the DSL allows this then we should support it.