GR #4116 Making holygradle bootstrap URL not be hard-coded in the init script.
This allows multiple sites to share source code, even if they don't share the
same server for the holygradle plugins.  The site-specific part of the URL now
goes into an environment variable.

* .hgignore
** Ignore files which are genreated during build (gw.bat) or bootstrapping
(distributionUrlBase.txt).

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\CustomGradleCorePlugin.groovy
** Extract the content of "gw.bat" and just copy it from the holygradle repo
root, to be be used as a resource from the JAR.
** When writing the new "gw.bat", write it to "gw.bat.new", so it can safely
self-upgrade (see below).
** Move the generated "gradle-wrapper.properties" to "gradle-wrapper.properties.in",
deleting the "distributionUrl" line.  This is used as input to "gw.bat" to paste
together a "gradle-wrapper.properties" which contains the site-specific part.
** Write the versioned distribution path to "distributionPath.txt", again for
use by "gw.bat".

* custom-gradle\src\initScripts\holy-gradle-init.gradle
** The plugin server name is no longer baked in.  Instead it's provided from an
enviroment variable, by "gw.bat".
** Added more comments to explain the new and existing behaviour.

* gw.bat
** Make this batch file self-upgrade itself from a "gw.bat.new" if it exists.
** Remove support for Win9x, WinME and 4NT, since we're doing too much fancy DOS
stuff, and holygradle requires Windows 7 anyway.
** Use the HOLY_GRADLE_REPOSITORY_BASE_URL environment variable, together with
the contents of "gradle\gradle-wrapper.properties.in" and
"gradle\distributionUrlBase.txt" to paste together a
"gradle\gradle-wrapper.properties" which has the full distributionUrl, with
site-specific part.

* custom-gradle-core-plugin\build.gradle
** Add a build step to copy the root project's "gw.bat" for use as a JAR resource.

* artifactory-manager-plugin\gw.bat
* credential-store\gw.bat
* custom-gradle-core-plugin\gw.bat
* custom-gradle\gw.bat
* devenv-plugin\gw.bat
* intrepid-plugin\gw.bat
* my-credentials-plugin\gw.bat
* stamper-plugin\gw.bat
* unit-test-plugin\gw.bat
** These are no longer used, so I removed them to avoid confusion with the
top-level batch file.