GR #4116 Making holygradle bootstrap URL not be hard-coded in the init script.
Reinstated the wrapper-starter-kit project, which makes a ZIP file of just the
files needed to us a given version of the holygradle custom distribution.

* .hgignore
** Update the ignores for the previously-removed, now-revived wrapper-starter-kit
project.
** Tidy up a few others in passing.

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\CustomGradleCorePlugin.groovy
** Remove unused import.

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\PrerequisitesExtension.groovy
** Fix bug uncovered by new wrapper-starter-kit project -- the
custom-gradle-core-plugin would fail to apply if the root project didn't apply
it (or a plugin which also applies it -- devenv or intrepid).

* developer-readme.md
** Added notes about setting up the wrapper-starter-kit project.

* gw.bat
** Updated to be the same as the one in the wrapper-starter-kit subproject.

* wrapper-starter-kit\build.gradle
** Project to make and publish a ZIP file of the files needed to use a given
holygradle custom-gradle distribution version.

* wrapper-starter-kit\gradle\gradle-wrapper.jar
* wrapper-starter-kit\gradle\gradle-wrapper.properties.in
** Gradle bootstrap files for the new project.  It has to be separate from the
root project because it _uses_ the published plugins, rather than building them.

* wrapper-starter-kit\gw.bat
** Modified to have special-case handling for bootstrapping when running this
project.  Otherwise you'd need to check in the distributionPath.txt file and
modify-then-commit it every time you wanted to publish a new starter kit version.
For diffs, compare the root-level gw.bat.

* wrapper-starter-kit\newWrapper\build.gradle
** Stub project to allow the custom-gradle-core-plugin's createWrapper task to
be used.  We don't want to createWrapper in the wrapper-starter-kit project, or
we'd overwrite the files we're using to publish.

* wrapper-starter-kit\settings.gradle
** Standard multi-project build settings file.