GR #3554 Allow more fine-grained deletion periods
[Just spotted that this got accidentally backed out as part of ae413040fa70,
because that included a back-out of merge 47d4ec8c75ae and hence its "other"
ancestors, which I actually wanted to keep.]

* artifactory-manager-plugin\src\main\groovy\holygradle\artifactory_manager\ArtifactoryManagerHandler.groovy
** Output list files to a sub-directory and, within that, to a timestamped
directory.  This makes it easier to compare size listings taken at different
times.

* artifactory-manager-plugin\src\main\groovy\holygradle\artifactory_manager\ArtifactoryManagerPlugin.groovy
** Update task description to include new output folder, and acknowledge that
there is more than one size file output per repo.

* artifactory-manager-plugin\src\main\groovy\holygradle\artifactory_manager\DeleteRequest.groovy
** Allow the "keep" interval size to be "day(s)", as well as weeks/months/years.
** Make sure that items dated exactly at midnight are treated sensibly when
working out whether or not they fall within the "sliding window" of versions
which are considered for keeping.
** When keeping one version in an interval, keep the most recent, not the
oldest, as it's more likely to be useful.

* artifactory-manager-plugin\src\main\groovy\holygradle\artifactory_manager\RepositoryHandler.groovy
** Dump out summaries per module and per version, as well as listing sizes down
to the level of individual files.

* artifactory-manager-plugin\src\main\groovy\holygradle\artifactory_manager\StorageSpaceLister.groovy
** Dump out summaries per module and per version.
** Make files TAB-separated, for quicker import into Excel.
** If what we expect to be a file doesn't have a size (because it's not there,
or because it's a folder), print an error in the summary but continue.  I've
seen this with some odd repositories where not all files deployed there fit the
intended repo pattern.

* artifactory-manager-plugin\src\test\groovy\holygradle\artifactory_manager\DeletionTest.groovy
** Added new test testDeleteKeepingOneItemPerDay.
** Update tests to match the fact that, if there are multiple versions in one
"keep" interval, we now keep the most recent, rather than the earliest.