GR #2584 Work around build_info race condition with a retry.
[Just spotted that this got accidentally backed out as part of ae413040fa70,
because that included a back-out of merge 47d4ec8c75ae and hence its "other"
ancestors, which I actually wanted to keep.]

Sometimes, deleting and creating the build_info folder fails, apparently due to
a race condition.  (I suspect the JDK's mkdir/rmdir methods are returning before
the file system is fully updated.)  This causes real problems from time to time,
so I'm working around it with a retry mechanism.