Changed SNAPSHOT override behaviour if using another plugin which depends on the holygradle plugins.
[Just spotted that this got accidentally backed out as part of ae413040fa70,
because that included a back-out of merge 47d4ec8c75ae and hence its "other"
ancestors, which I actually wanted to keep.]

Changed SNAPSHOT override behaviour if using another plugin which depends on the
holygradle plugins.  If the global override was used, there was no problem; but
if only some plugins (e.g., all the core holygradle plugins) were set to version
"username-SNAPSHOT", Gradle would follow default version resolution rules,
which might result in the SNAPSHOT override version not being used, if that
other plugin depended on a non-SNAPSHOT version.  Now it will always force the
SNAPSHOT version to win, for any plugin specified as using a snapshot version.

This makes it possible to patch in a SNAPSHOT, e.g., with extra logging, to
investigate the behaviour of non-SNAPSHOT plugin versions.