Fix bugs in publishing, due to my misunderstanding of how dependencies and tasks work.
* holy-gradle-plugins\credential-store\build.gradle
* holy-gradle-plugins\custom-gradle\build.gradle
** If you want one task, B, to use the outputs of another, A, it's not enough to
have B depend on A; you need to add A's outputs as inputs of B.  Otherwise,
whenever B is put in the task graph (scheduled to run), A will also be BUT even
if A actually runs, B may not (unless its outputs are missing).  (I should
already know that, but most of the tasks I've written haven't had that kind of
dependency.)

* holy-gradle-plugins\publish.gradle
* holy-gradle-plugins\versionInfo.gradle
** While investigating this problem, I noticed that tasks setPublishVersion and
setHgVersionInfo were writing their output files at configuration time, rather
than at execution time.  That may or may not have been contributing to the
problem, but I fixed it anyway.