GR #4371 Put source back references into gw.bat and https-setup-kit.
For wrapper-starter-kit, the version info now in gw.bat will tell us where the
wrapper came from (and we don't want any extra files in that wrapper, to clutter
up source repositories of new projects).

For https-setup-kit we use a plain Gradle wrapper, so we copy in a separate file
to tell us.

* custom-gradle-core-plugin\build.gradle
* gw.bat
** Add version info to gw.bat when building.

* https-setup-kit\build.gradle
** Add version info file to the ZIP when building.