GR #3813, GR #4071 Changing how ivy.xml files are found whlie visiting unpack module dependencies, so that intrepid can traverse through configurations which have no artifacts.
* .idea\dynamic.xml
** Add IntelliJ dynamic property info for new extension.

* intrepid-plugin\src\main\groovy\holygradle\IntrepidPlugin.groovy
** Move the unpackModule info into a Gradle-style extension, instead of having
it just be static state on some class.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\UnpackModule.groovy
** Move the dependency traversal code to new PackedDependenciesStateHandler
class, and refactor it in the process.
** Add a toString override.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\UnpackModuleVersion.groovy
** Add a toString override.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\PackedDependenciesStateHandler.groovy
** The existing method for finding ivy files is as follows, which has a few
problems.  First, it can pick up the wrong ivy.xml file if there's more than one
in the Gradle cache folder hierarchy, which can happen if the module has been
downloaded more than once with different contents.  Second, it fails to follow
dependencies through configurations which have no artifacts, because it finds
the ivy.xml in terms of a relative path from the first artifact.
*** Traverse the full set of ResolvedDependencies, for a given configuration.
*** For each one, try to find the ivy file by (a) finding the location of the
first artifact in the user's Gradle cache; (b) searching reursively for the
first "ivy.xml" file in a directory hierarchy relative to that file.
** The new mechanism is as follows.
*** Resolve all dependencies as normal.
*** Create an entirely new configuration at runtime (and make sure it's not
published).
*** Traverse the full set of ResolvedDependencies for a given configuration and,
for each module which isn't part of the current build, add a new (non-resolved)
dependency to the new configuration.  (These are added as direct dependencies --
we don't need to mirror the original dependency graph structure, just get all
the deps.)
*** For each such dependency, add just one artifact, the Ivy file.
*** Now resolve this new configuration, and pull out the Ivy file names.