GR #3813, GR #4071 Fixed bug that only the first configuration per dependency-on-a-module was visited.  (This bug was introduced as part of my refactoring to fix this bug, I believe, though I'm not sure since there were no tests before I started.)
* .hgignore
** Ignore a few files I should have ignored a while ago, with version info, plus
updated one for a rename in a recent commit.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\PackedDependenciesStateHandler.groovy
** When visiting configurations in the original project's dependency graph (as
opposed to in the detached configurations for ivy.xml files), we need to track
not just which "module versions" we've seen, but which "module version plus
module configuration", since dependencies differ per configuration.
** Turned logger.info calls into debug.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\UnpackModuleVersion.groovy
** Adjusted name and doc for a field, when I realised that the configurations
named here are not from the module to which this version applies, but from the
original entry project.

* intrepid-plugin\src\test\groovy\holygradle\testUtil\repository\holygradle.test\another-lib\1.0\ivy-1.0.xml
* intrepid-plugin\src\test\groovy\holygradle\testUtil\repository\holygradle.test\empty-config-lib\1.0\ivy-1.0.xml
** Fix the names of artifacts (broken since I copy'n'pasted these ivy files from
another, rather than getting Gradle to create them).  Otherwise, the plugins
complain about modules/configurations with no artifacts.

* intrepid-plugin\src\test\groovy\holygradle\unpacking\PackedDependenciesStateHandlerIntegrationTest\useModulesViaEmptyConfigs\build.gradle
** Add a dependency on a module configuration which has an artifact, otherwise
the plugins complain that some module has no artifacts.  (So far we treat that
as probably a user error, though I may change it to be allowed.)