GR #3813, GR #4071 Don't visit the children of dependencies in the ivy.xml configurations.
When finding ivy files, we generate a "detached configuration" which just has
direct dependencies on all the module versions in the original dependency graph,
with the artifact list replaced with one which just has the ivy file.  We then
visit all modules in *that* configuration to resolve the ivy file locations.
When doing that latter step, we only need to visit the first-level dependencies
amd, indeed, we *must* only visit those.  If we visit transitive ones, we will
find the original dependencies, which don't have the ivy files.  This may or may
not cause a problem, depending on the order in which we visit them since we don't
re-visit ones whose "group:name:version" we've already seen -- if we visit the
top-level one first, we get the ivy file; if not, we don't.

So, I've turned off visiting of children in this case, and added a double-check
at the point where we should have found the ivy file.