GR #4088 If a module has no artifacts at all, collectDependencies won't collect it.
* intrepid-plugin\src\main\groovy\holygradle\dependencies\CollectDependenciesTask.groovy
** Refactored extensively.  Information on the location of Ivy XML and Maven POM
files is now taken from instance of the new DependenciesStateHandler class,
rather than working backward from the resolved artifact paths in the cache,
using relative paths plus "traverse" -- that's not guaranteed to work, because
the same metadata file may exist in more than one version, if it was different
at different times or on different servers.
** Also cleaned up copying of the custom-gradle sitribution and rewriting of the
gradle-wrapper.properties file, to use specific filenames rather than searching
for "likely-looking" files.

* intrepid-plugin\src\main\groovy\holygradle\IntrepidPlugin.groovy
** Adding two instances of the new DependenciesStateHandler class as extensions.
In the process, corrected comments about extensions relating to "the build
script", when they really mean "the project" (since I now have an extension
relating to "the [project's] buildscript").

* intrepid-plugin\src\main\groovy\holygradle\unpacking\PackedDependenciesStateHandler.groovy
** Code which relates to dependencies in general, rather than the unpacking of
packed dependencies, was split off to new classes DependenciesStateHandler and
ResolvedDependenciesVisitor, so it could be shared with CollectDependenciesTask.
** Added a big explanatory comment at the top, to capture the meaning I
eventually deciphered in this class.
** Other minor tidying.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\UnpackModule.groovy
** Added deprecation warning.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\DependenciesStateHandler.groovy
** Factored out from PackedDependenciesStateHandler and extended to cope with
Maven dependencies (including parent/ancestor POMs), and dependencies of a
project's buildscript, as separate from those of the project itself.
** Added lots of doc.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\ResolvedDependenciesVisitor.groovy
** Factored out from PackedDependenciesStateHandler so it could be used by both
that class and the new DependenciesStateHandler.