GR #4546 addPackedDependency is broken in buildScript handler
Was using 'wildcard' matching to search the project for the relevant source/
packed dependency: the documentation clearly states you should use the name
of the dependency without the path.  So have modified the behaviour to comply.

Extended PackageArtifactBuildScriptHandlerIntegrationTest which fails (and
replicates the defect) without the corresponding change to
PackageArtifactBuildScriptHandler.