GR #4546 addPackedDependency is broken in buildScript handler
Ok so the wiki documentation states that for addPinnedSource, 'wildcard' matching
is used.  And this handler uses the same method (findSourceDependencies) as
addPackedDependency.

So this commit:

* fixes the Wildcard.match method (in custom_gradle.util) to
do a safer glob match with help from regex (i.e. so that foo will not match
foobar, but foo* will match foobar)

* adds unit-test for above

* keeps the added logic for findSourceDependencies from the previous commit but
allow allows this method to accept wildcards (to be used in the addPinnedSource
case)