GR #4125 Quick attempt to improve holygradle start-up time
This is an attempt to speed up start-up time for the holygradle plugins on large
projects, by avoiding calls to getAllUnpackModules unless specific tasks are
being executed.  Calling that method resolves all configurations, which can be
very slow, specially now that we're resolving an extra set of configurations to
correctly get ivy.xml files.

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\CustomGradleCorePlugin.groovy
** Add profiling of plugin startup time.

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\util\Symlink.groovy
** Don't launch mklink.exe to create symlinks, just do it from Java.  This means
that we can't create directory symlinks to directories which don't exist yet --
they're created as file symlinks -- but we've never actually needed to do that.
(This also means the fileLink argument isn't needed, since Java decides based on
the type of the target.)

* devenv-plugin\src\main\groovy\holygradle\devenv\DevEnvPlugin.groovy
** Add profiling of plugin startup time.

* integration\src\test\groovy\holygradle\BasicIntegrationTest_tAPIT_stdout.ok.txt
* integration\src\test\groovy\holygradle\BasicIntegrationTest_tBPC_stdout.ok.txt
** All symlinks-to-cache are now created/deleted by one task.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\PackedDependencyHandler.groovy
** Add method noCreateSymlinkToCache().  This explicit option replaces the
previous implicit and sometimes confusing behaviour of SymlinkTask -- if a file
already existed at the link (not target) location, it would print a message and
continue, rather than fail the build.  Allowing this can be useful if you want
to temporarily swap in a personal build of a packedDependency, but doing it
automatically has confused users several times.  If you don't call this method,
then a symlink will be created if unpackToCache is true, and not if it's false.
** In passing, deprecate unneeded and unused extra syntax "unpackToCache false";
use "unpackToCache = false".

* intrepid-plugin\src\main\groovy\holygradle\IntrepidPlugin.groovy
** Add profiling of plugin startup time, and of various internal sections.
** Use new SymlinksToCacheTask for creating/deleting symlinks to the unpack
cache.  This is instead of creating one task per packed dependency, because
doing that requires us to call getAllUnpackModules when we're creating the list
of tasks (i.e., on every build), rather than just when we need to update
symlinks.
*** Make this a dependency of the existing "delete/rebuildSymlinks" tasks.
** Previously we called getAllUnpackModules to create one unpack task per
packed dependency.  I didn't want to combine all these into one task, as I did
with the "symlinks to cache", because having them separate lets us use Gradle's
usual task up-to-date check to see whether we need to unpack things.  Instead,
we now only create those tasks if "fetchAllDependencies" is specified as a
target on the command line.  This means you can't make another task which
"dependsOn" it and just use that as a target, but nobody does that anyway.
** createSettingsFileForPackedDependencies and checkPackedDependencies tasks
also now delays calling getAllUnpackModules until they're actually executed.

* intrepid-plugin\src\main\groovy\holygradle\publishing\CheckPublishedDependenciesTask.groovy
* intrepid-plugin\src\main\groovy\holygradle\publishing\DefaultPublishPackagesExtension.groovy
* intrepid-plugin\src\main\groovy\holygradle\publishing\PublishPackagesExtension.groovy
** Delay calling getAllUnpackModules until the task is actually executed.

* intrepid-plugin\src\main\groovy\holygradle\source_dependencies\RecursivelyFetchSourceTask.groovy
** Added a "mark as initialized" method which is called by IntrepidPlugin when
this task is specified on the command line.  This is so that, if anyone tries to
run it indirectly, as a dependency of another task, it fails with a useful
message.

* intrepid-plugin\src\main\groovy\holygradle\symlinks\SymlinkTask.groovy
** If a link source already exists, this task now fails, instead of just
printing a message and continuing.
** Also, we now keep a map of requested symlinks and iterate over that within a
single doFirst block, rather than adding one doFirst block per symlink; that's
got to be faster, surely?!?
** Also renamed the unclear "configure" to "addLink".

* intrepid-plugin\src\main\groovy\holygradle\unpacking\PackedDependenciesStateHandler.groovy
** Factor out a minimal interface so we can pass this handler to things (tasks)
which only need the results.
** Added missing return type.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\UnpackModuleVersion.groovy
** The task code for creating symlinks to the unpackCache has moved to the new
class SymlinksToCacheTask.  There's some back-and-forth calling between the UMV
and the task to find out whether each UMV actually needs/wants a symlink.
** In d723b077f1c6 I made the constructor treat relateiPaths with a trailing '/'
differently.  I should have done the same in getTargetPathInWorkspace (so now I
have).
** Renamed getParentPackedDependency to getSelfOrAncestorPackedDependency,
because that's what it actually does.
** Added shouldCreateSymlinkToCache, to support
PackedDependencyHandler#noCreateSymlinkToCache().
** In passing, put more into the toString(), for debugging.

* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest\projectC\build.gradle
** This test relied on the old SymlinkTask behaviour of continuing when a
symlink couldn't be created, so I've made that explicit now.

* intrepid-plugin\src\test\groovy\holygradle\unpacking\UnpackModuleVersionTest.groovy
** Renamed getParentPackedDependency to getSelfOrAncestorPackedDependency.

* my-credentials-plugin\src\main\groovy\holygradle\credentials\MyCredentialsPlugin.groovy
** Add profiling of plugin startup time.

* stamper-plugin\src\main\groovy\holygradle\stamper\StamperPlugin.groovy
** Add profiling of plugin startup time, and of various internal sections.

* unit-test-plugin\src\main\groovy\holygradle\unit_test\UnitTestPlugin.groovy
** Add profiling of plugin startup time, and of various internal sections.

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\util\ProfilingHelper.groovy
** Class to do some simple logging of elapsed time in named blocks of code.

* intrepid-plugin\src\main\groovy\holygradle\symlinks\SymlinksToCacheTask.groovy
** New class which encapsulates creating symlinks into the unpackCache for all
UnpackModuleVersions (packedDependencies and their transitive dependencies).

* intrepid-plugin\src\main\groovy\holygradle\unpacking\PackedDependenciesStateSource.groovy
** Factored out from PackedDependenciesStateHandler.